% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentWorkflowFunctions.R
\name{addPvaluesToAllezOutput}
\alias{addPvaluesToAllezOutput}
\title{Helper function to add p-values to the final output of the Allez workflow}
\usage{
addPvaluesToAllezOutput(
  outputAllez,
  Lowersetsize = 5,
  Uppersetsize = 500,
  side = "T"
)
}
\arguments{
\item{outputAllez}{The object created at the end of the EnrichAllez function}

\item{Lowersetsize}{terms with less than this many instances are dropped from the list}

\item{Uppersetsize}{Terms with greater than this many instances are dropped from the list}

\item{side}{If True, indicates that the p-values distribution is two-tailed}
}
\value{
Combines the original outputAllez object with the p-values
}
\description{
Helper function to add p-values to the final output of the Allez workflow
}
