% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentWorkflowFunctions.R
\name{convertAccessions}
\alias{convertAccessions}
\title{Converts uniprot accessions via biomaRt}
\usage{
convertAccessions(
  wgcnaResults,
  convertTo = "external_gene_name",
  filterBy = "uniprot_gn_id",
  mart
)
}
\arguments{
\item{wgcnaResults}{The complete results from the WGCNA portion of the shiny app}

\item{convertTo}{What the accessions will be converted to. Defaults to gene symbol}

\item{filterBy}{What the accession currently are. Default is UniProt accessions}

\item{mart}{Object of type biomart}
}
\value{
A list of length(wgcnaResults) where each list contains the gene symbols
converted from UniProt accessions
@name convertAccessions
@export
}
\description{
Converts uniprot accessions via biomaRt
}
