% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAFunctions.R
\name{createResultsWGCNAExcelWorkbook}
\alias{createResultsWGCNAExcelWorkbook}
\title{Write WGCNA Excel Output}
\usage{
createResultsWGCNAExcelWorkbook(
  fullDataSet,
  modulesList,
  filePath = "Results/ResultsWGCNA.xlsx"
)
}
\arguments{
\item{fullDataSet}{Object including all data for each accession.}

\item{modulesList}{List, with each element in the list corresponding to a single module.}

\item{filePath}{The file to ultimately write to}
}
\description{
Write WGCNA Excel Output
}
