% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentWorkflowFunctions.R
\name{fetchMart}
\alias{fetchMart}
\title{Gets the correct mart from biomaRt}
\usage{
fetchMart(dataSetIndex)
}
\arguments{
\item{dataSetIndex}{An index input by the shiny reactive values.}
}
\value{
An object of type "mart".
}
\description{
Gets the correct mart from biomaRt
}
