% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAFunctions.R
\name{mergeModuleEigenproteins}
\alias{mergeModuleEigenproteins}
\title{Merges correlatd module eigenproteins}
\usage{
mergeModuleEigenproteins(exprData, moduleColors, cutHeight)
}
\arguments{
\item{exprData}{Matrix of expression data}

\item{moduleColors}{Module membership for each protein}

\item{cutHeight}{User input value corresponding to the desired merge height.}
}
\value{
Merged module membership and hclust objects
}
\description{
Merges correlatd module eigenproteins
}
