% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAFunctions.R
\name{plotAdjacencyHeatmap}
\alias{plotAdjacencyHeatmap}
\title{Plots module adjacency as a heatmap.}
\usage{
plotAdjacencyHeatmap(
  moduleEigenproteins,
  fileName = "Results/Eigenprotein_adjacency heatmap.pdf",
  widthInches = 10
)
}
\arguments{
\item{moduleEigenproteins}{Datatable of module eigenproteins and their corresponding values for each sample.}

\item{fileName}{Desired filename.}

\item{widthInches}{Set the width of the output. Deafult parameter is
10 inches.}
}
\value{
Creates a plot of the adjacency heatmap, which relates the adjacency between the
module eigenproteins. Hierarchical clustering is performed using 1-adjacency,
where adjacency is calculated by transforming the correlation. Can show
interactions between two modules.
@name plotAdjacencyHeatmap
@export
}
\description{
Plots module adjacency as a heatmap.
}
