% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAFunctions.R
\name{plotEigenproteinClusteringPostMerge}
\alias{plotEigenproteinClusteringPostMerge}
\title{plots the eigenprotein clustering after module merging}
\usage{
plotEigenproteinClusteringPostMerge(
  preMergeDendroData,
  postMergeDendroData,
  cutHeight,
  fileName = "Results/ModuleEigenproteinMergeDendrogram.pdf",
  widthInches = 10
)
}
\arguments{
\item{preMergeDendroData}{dendrogram of module eigenproteins before module merging}

\item{postMergeDendroData}{dendrogram of module eigenprotiens after module merging}

\item{cutHeight}{the user input module merge cut height}

\item{widthInches}{Default to 10 inches}

\item{filenName}{file name for plot saving}
}
\value{
Outputs pdf file of the denrograms of the eigenproteins both before
and after module merging.
}
\description{
plots the eigenprotein clustering after module merging
}
