% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAFunctions.R
\name{plotEigenproteinsNetwork}
\alias{plotEigenproteinsNetwork}
\title{Plots the correlation-based eigenprotein dendrogram}
\usage{
plotEigenproteinsNetwork(
  moduleEigenproteins,
  fileName = "Results/DendrogramEigenproteins.pdf",
  widthInches = 10
)
}
\arguments{
\item{moduleEigenproteins}{Datatable of module eigenproteins and their corresponding values for each sample.}

\item{fileName}{Desired filename.}

\item{widthInches}{Set the width of the output. Deafult parameter is 10 inches.}
}
\value{
Creates a plot of the eigenproteins based on correlation dissimilarity (1-correlation(module_i, module_j)).
}
\description{
Plots the correlation-based eigenprotein dendrogram
}
