% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAFunctions.R
\name{plotModuleEigenproteinsBySample}
\alias{plotModuleEigenproteinsBySample}
\title{Plots module eigenproteins by sample}
\usage{
plotModuleEigenproteinsBySample(MEPs, module, filename)
}
\arguments{
\item{MEPs}{Module eigenproteins object with}

\item{module}{Color-based module identification}

\item{filename}{The desired filename to save each plot to}
}
\value{
Plots the module eigenproteins by sample
}
\description{
Plots module eigenproteins by sample
}
