% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAFunctions.R
\name{plotmoduleEigenproteinsHeatmap}
\alias{plotmoduleEigenproteinsHeatmap}
\title{Plots samples versus module eigenproteins in heatmap form.}
\usage{
plotmoduleEigenproteinsHeatmap(
  moduleEigenproteins,
  fileName = "Results/ModuleEigenproteinsHeatmap.pdf",
  widthInches = 15
)
}
\arguments{
\item{moduleEigenproteins}{Datatable of module eigenproteins and their
corresponding values for each sample.}

\item{fileName}{Desired filename.}

\item{widthInches}{Set the width of the output. Deafult parameter is 10 inches.}
}
\value{
Creates a heatmap that shows samples vs eigenprotein, with coloration
dependent on the value of the eigenprotein.
}
\description{
Plots samples versus module eigenproteins in heatmap form.
}
