% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAFunctions.R
\name{plotProteinDendrogram}
\alias{plotProteinDendrogram}
\title{Plots protein dendrogram with color bar identifying module membership.}
\usage{
plotProteinDendrogram(proteinDendrogram, mergedColors)
}
\arguments{
\item{proteinDendrogram}{Object of type hclust containing the clustering results.}

\item{mergedColors}{Vector of the module membership of each protein in the
proteinDendrogram.}
}
\value{
Creates a hierarchical clustering dendrogram with the module
membership of each protein beneath.
#' @importFrom grDevices pdf
}
\description{
Plots protein dendrogram with color bar identifying module membership.
}
