% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAFunctions.R
\name{plotSampleClusteringDenro}
\alias{plotSampleClusteringDenro}
\alias{plotSampleClusteringDendro}
\title{Plots correlation-based sample clustering.}
\usage{
plotSampleClusteringDendro(
  dendro,
  fontSize = 15,
  fileName = "SampleClustering.pdf"
)
}
\arguments{
\item{dendro}{Dendrogram of samples based on dissimlarity correlation.}

\item{fontSize}{Default = 15.}

\item{fileName}{Preferred file name.}
}
\value{
Creates a plot of the hierarchical clustering of the samples,
which can identify sample outliers.
}
\description{
Plots correlation-based sample clustering.
}
