% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAFunctions.R
\name{plotScaleFreeTopology}
\alias{plotScaleFreeTopology}
\title{Plots scale free topology diagnostic plots.}
\usage{
plotScaleFreeTopology(
  fileName = "Results/ScaleFreeTopology.pdf",
  powersOutput,
  powers,
  cex1 = 0.9,
  scaleFreeThreshold,
  widthInches = 10
)
}
\arguments{
\item{fileName}{preferred file name.}

\item{powersOutput}{Result of WGCNA::picksfotthreshhold.}

\item{powers}{Vector of powers from 1:user input value}

\item{cex1}{Resizes the final graphics device}

\item{scaleFreeThreshold}{User input value corresponding to the R^2 topology threshold}

\item{widthInches}{Default size of the graphical output}
}
\value{
Creates the diagnostic plot showing the satisfaction of the scale-free
topology and the effect of power on the mean connectivity of the proteins.
#' @importFrom grDevices pdf
}
\description{
Plots scale free topology diagnostic plots.
}
