% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAFunctions.R
\name{plotTOM}
\alias{plotTOM}
\title{Plots the topological overlap matrix as a heatmap.}
\usage{
plotTOM(
  TOMData,
  proteinDendro,
  moduleColors,
  fileName = "Results/NetworkHeatmap.png",
  ...
)
}
\arguments{
\item{TOMData}{TOM object output by the WGCNA workflow}

\item{proteinDendro}{TOM-based hclust object from workflow}

\item{moduleColors}{Module colors post-merging}

\item{fileName}{desired file name}

\item{...}{passes arguments to TOMplot function}
}
\value{
Outputs the network heatmap.
}
\description{
Plots the topological overlap matrix as a heatmap.
}
