% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAFunctions.R
\name{plotTOMpdf}
\alias{plotTOMpdf}
\title{Plots the topological overlap matrix as a heatmap. Saves as a pdf.}
\usage{
plotTOMpdf(
  TOMData,
  proteinDendro,
  moduleColors,
  fileName = "Results/NetworkHeatmap.pdf",
  height = 12,
  width = 12
)
}
\arguments{
\item{TOMData}{TOM object output by the WGCNA workflow}

\item{proteinDendro}{TOM-based hclust object from workflow}

\item{moduleColors}{Module colors post-merging}

\item{fileName}{desired file name}

\item{height}{height of pdf output}

\item{width}{width of pdf output}
}
\value{
Outputs the network heatmap as a pdf. Use only for small TOMs
}
\description{
Plots the topological overlap matrix as a heatmap. Saves as a pdf.
}
