% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{read.csv_override}
\alias{read.csv_override}
\title{Override for read.csv that sets stringsAsFactors to TRUE.}
\usage{
read.csv_override(fileName, ...)
}
\arguments{
\item{fileName}{default file name to pass to results folder.}

\item{...}{Params passed to read.csv. See documentation for read.csv for full list of parameters}
}
\value{
override for read.csv that sets stringsAsFactors = TRUE by default. Will prevent future errors
}
\description{
Override for read.csv that sets stringsAsFactors to TRUE.
}
