% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Reads an excel sheet and imports each sheet as a component of a list.}
\usage{
read_excel_allsheets(filename, tibble = FALSE)
}
\arguments{
\item{filename}{path to excel sheet to be loaded}

\item{tibble}{make the data imported a tibble. Default = FALSE. Data imported as a data.frame.}
}
\value{
List where each list elements corresponds to an excel sheet.
}
\description{
Reads an excel sheet and imports each sheet as a component of a list.
}
