% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallApp.R
\name{RunMetaNetwork}
\alias{RunMetaNetwork}
\title{Starts MetaNetwork}
\usage{
RunMetaNetwork(results_directory)
}
\arguments{
\item{...}{Arguments to pass to shiny::runApp}
}
\value{
Runs the MetaNetwork shiny application
}
\description{
Function to run MetaNetwork shiny application
}
