% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAFunctions.R
\name{writeVarianceExplained}
\alias{writeVarianceExplained}
\title{Writes a .csv file of the variance explained by the module eigenproteins}
\usage{
writeVarianceExplained(datExpr, colors, MEs)
}
\arguments{
\item{datExpr}{Samples vs proteins data frame}

\item{colors}{module colors}

\item{MEs}{module eigenproteins}
}
\value{
Write a .csv file that shows how well the module
eigenproteins explain the variance in the data.
}
\description{
Writes a .csv file of the variance explained by the module eigenproteins
}
